backend-flask/requirements.txt
flask
flask-cors
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp-proto-http 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-requests
aws-xray-sdk
watchtower
blinker
rollbar

Flask-AWSCognito

psycopg[binary]
psycopg[pool]

boto3
pytest

stripe

momento==1.7.0
momento-wire-types==0.67.0